--- a/support/shake/app/Shake/Digest.hs
+++ b/support/shake/app/Shake/Digest.hs
@@ -2,7 +2,10 @@
 
 -- | Compute a truncated hash of a file, useful for computing cache-busters
 -- (or other unique ids) for a file.
-module Shake.Digest (digestRules, getFileDigest) where
+module Shake.Digest (shortDigest, digestRules, getFileDigest) where
+
+import qualified Data.Text.Encoding as Text
+import qualified Data.Text as Text
 
 import qualified Data.ByteString.Lazy as LazyBS
 import Data.Digest.Pure.SHA
@@ -27,3 +30,6 @@ digestRules = versioned 1 do
 -- | Compute a short digest of a file.
 getFileDigest :: FilePath -> Action String
 getFileDigest = askOracle . FileDigest
+
+shortDigest :: Text.Text -> [Char]
+shortDigest = take 12 . showDigest . sha1 . LazyBS.fromStrict . Text.encodeUtf8