--- a/support/web/css/code.scss
+++ b/support/web/css/code.scss
@@ -2,6 +2,7 @@
 
 @use "vars.scss"   as *;
 @use "mixins.scss" as *;
+@use "theme.scss"  as *;
 
 @font-face {
   font-family: 'Julia Mono';