… .github/workflows/build.yml
… support/nix/dep/Agda/github.json
… support/shake/app/HTML/Emit.hs
… support/shake/app/Main.hs
… support/shake/app/Shake/AgdaCompile.hs
… support/shake/app/Shake/Diagram.hs
… support/shake/app/Shake/KaTeX.hs
… support/shake/app/Shake/LinkGraph.hs
… support/shake/app/Shake/LinkReferences.hs
… support/shake/app/Shake/Markdown.hs
… support/shake/app/Shake/Markdown.hs-boot
… support/shake/app/Shake/Markdown/Reader.hs
… support/shake/app/Shake/Modules.hs
… support/shake/app/Shake/Options.hs
… support/shake/app/Shake/Recent.hs
… support/web/css/components/commit.scss
… support/web/css/components/controls.scss
… support/web/css/components/modal.scss
… support/web/css/components/popup.scss
… support/web/css/components/search.scss
… support/web/css/components/sidenotes.scss
… support/web/css/components/toc.scss
… support/web/css/default.scss
… support/web/css/mixins.scss
… support/web/css/talk.scss
… support/web/css/theme.scss
… support/web/js/highlight-hover.ts
… support/web/js/theme.tsx
… support/web/template.html
… support/web/template.reveal.html