Time to do a shake build from a clean database and build directory but with everything typechecked on `main`:

```
 574,934,516,384 bytes allocated in the heap
  32,522,090,000 bytes copied during GC
   7,503,307,808 bytes maximum residency (9 sample(s))
     573,250,528 bytes maximum slop
           19446 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       181 colls,   181 par   152.304s   9.818s     0.0542s    0.2426s
  Gen  1         9 colls,     8 par   216.985s   9.099s     1.0110s    4.3868s

  Parallel GC work balance: 86.52% (serial 0%, perfect 100%)

  TASKS: 111 (1 bound, 110 peak workers (110 total), using -N32)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.002s  (  0.002s elapsed)
  MUT     time  693.233s  ( 69.977s elapsed)
  GC      time  369.290s  ( 18.917s elapsed)
  EXIT    time    0.811s  (  0.004s elapsed)
  Total   time  1063.335s  ( 88.900s elapsed)

  Alloc rate    829,352,576 bytes per MUT second

  Productivity  65.2% of total user, 78.7% of total elapsed
```

Rubbish. On this branch:

```
 399,468,065,608 bytes allocated in the heap
   9,622,615,480 bytes copied during GC
   3,395,392,848 bytes maximum residency (7 sample(s))
      21,588,656 bytes maximum slop
           10561 MiB total memory in use (0 MiB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       121 colls,   121 par   75.455s   3.977s     0.0329s    0.2443s
  Gen  1         7 colls,     6 par    7.335s   0.588s     0.0840s    0.2552s

  Parallel GC work balance: 88.90% (serial 0%, perfect 100%)

  TASKS: 105 (1 bound, 104 peak workers (104 total), using -N32)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.053s  (  0.049s elapsed)
  MUT     time  466.568s  ( 54.406s elapsed)
  GC      time   82.790s  (  4.565s elapsed)
  EXIT    time    0.256s  (  0.002s elapsed)
  Total   time  549.667s  ( 59.022s elapsed)

  Alloc rate    856,183,933 bytes per MUT second

  Productivity  84.9% of total user, 92.2% of total elapsed
```