@plt-amy 2025-08-28
That memory usage is still... extremely high, but I did spend a while looking for GC parameters that shrink the heap to prevent CI getting exploded. Here's what I pushed (`-F1.25 -A8M`):

```
 399,515,229,016 bytes allocated in the heap
  53,406,314,032 bytes copied during GC
   3,688,678,440 bytes maximum residency (63 sample(s))
      40,413,776 bytes maximum slop
            5707 MiB total memory in use (0 MiB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     12351 colls, 12351 par   130.430s  22.358s     0.0018s    0.0286s
  Gen  1        63 colls,    62 par   85.880s   4.526s     0.0718s    0.2402s

  Parallel GC work balance: 59.60% (serial 0%, perfect 100%)

  TASKS: 105 (1 bound, 104 peak workers (104 total), using -N32)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.009s  (  0.006s elapsed)
  MUT     time  491.500s  ( 51.516s elapsed)
  GC      time  216.309s  ( 26.884s elapsed)
  EXIT    time    0.247s  (  0.002s elapsed)
  Total   time  708.065s  ( 78.407s elapsed)

  Alloc rate    812,849,685 bytes per MUT second

  Productivity  69.4% of total user, 65.7% of total elapsed
```

This *halves* the OS-memory usage at the cost of a ~33% increase in runtime. But perhaps `-F1.3 -A32M` is a better trade-off:

```
 399,454,760,808 bytes allocated in the heap
  24,812,172,880 bytes copied during GC
   3,640,078,296 bytes maximum residency (28 sample(s))
      33,646,152 bytes maximum slop
            6474 MiB total memory in use (0 MiB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       513 colls,   513 par   122.590s   6.585s     0.0128s    0.0930s
  Gen  1        28 colls,    27 par   39.841s   2.205s     0.0788s    0.2841s

  Parallel GC work balance: 85.65% (serial 0%, perfect 100%)

  TASKS: 107 (1 bound, 106 peak workers (106 total), using -N32)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.021s  (  0.016s elapsed)
  MUT     time  484.486s  ( 54.541s elapsed)
  GC      time  162.431s  (  8.790s elapsed)
  EXIT    time    0.221s  (  0.002s elapsed)
  Total   time  647.159s  ( 63.348s elapsed)

  Alloc rate    824,491,218 bytes per MUT second

  Productivity  74.9% of total user, 86.1% of total elapsed
```